made flourish variadic and accept lists like arithmetic stratagems
